* new Extension API.
* change Strategy interface.

* add middleware function.
* remove Typeable restriction from Path/Query class.
* remove unused Functor restriction.

* add Optional Strategy, (=?!:) query fetcher.
* add accept filter.
* add Path/Query instances to Day.

clientsession: add sessionTimeoutAction option

+++ .travis-install.sh
@@ -1,10 +1,17 @@
 #!/usr/bin/env bash -eu
 
-path=(".")
+path=("." "./examples")
 for p in `cat submodules`; do
   path=("${path[@]}" "./$p")
 done
 
 cabal --version
 cabal sandbox init
-cabal install --force-reinstalls --only-dependencies --enable-tests "${path[@]}"
+
+if [ -e restriction ]; then
+  cabal install --force-reinstalls --only-dependencies --enable-tests "${path[@]}" `cat restriction`
+  cabal install --force-reinstalls `cat restriction`
+else
+  cabal install --force-reinstalls --only-dependencies --enable-tests "${path[@]}"
+  cabal install --force-reinstalls 
+fi
